Move sds ingress tests to a single suite
