+++ tests/integration/security/sds_ingress/multiple_mtls_gateway_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package multipletlsgateway
+package sdsingress
 
 import (
 	"testing"
@@ -24,38 +24,36 @@ import (
 	ingressutil "istio.io/istio/tests/integration/security/sds_ingress/util"
 )
 
-var (
-	credNames = []string{"bookinfo-credential-1", "bookinfo-credential-2", "bookinfo-credential-3",
-		"bookinfo-credential-4", "bookinfo-credential-5"}
-	hosts = []string{"bookinfo1.example.com", "bookinfo2.example.com", "bookinfo3.example.com",
-		"bookinfo4.example.com", "bookinfo5.example.com"}
-)
-
-// testMultiTlsGateways deploys multiple TLS gateways with SDS enabled, and creates kubernetes that store
-// private key and server certificate for each TLS gateway. Verifies that all gateways are able to terminate
-// SSL connections successfully.
-func testMultiTLSGateways(t *testing.T, ctx framework.TestContext) { // nolint:interfacer
+// testMultiMtlsGateways deploys multiple mTLS gateways with SDS enabled, and creates kubernetes that store
+// private key, server certificate and CA certificate for each mTLS gateway. Verifies that all gateways are able to terminate
+// mTLS connections successfully.
+func testMultiMtlsGateways(t *testing.T, ctx framework.TestContext) { // nolint:interfacer
 	t.Helper()
 
-	ingressutil.CreateIngressKubeSecret(t, ctx, credNames, ingress.TLS, ingressutil.IngressCredentialA)
-	ingressutil.DeployBookinfo(t, ctx, g, ingressutil.MultiTLSGateway)
+	ingressutil.CreateIngressKubeSecret(t, ctx, multipleTLSCredNames, ingress.Mtls, ingressutil.IngressCredentialA)
+	ingressutil.DeployBookinfo(t, ctx, g, ingressutil.MultiMTLSGateway)
 
-	ing := ingress.NewOrFail(t, ctx, ingress.Config{Istio: inst})
-	err := ingressutil.WaitUntilGatewaySdsStatsGE(t, ing, len(credNames), 30*time.Second)
+	ing := ingress.NewOrFail(t, ctx, ingress.Config{
+		Istio: inst,
+	})
+	// Expect 2 SDS updates for each listener, one for server key/cert, and one for CA cert.
+	err := ingressutil.WaitUntilGatewaySdsStatsGE(t, ing, 2*len(multipleTLSCredNames), 30*time.Second)
 	if err != nil {
 		t.Errorf("sds update stats does not match: %v", err)
 	}
-	// Expect two active listeners, one listens on 443 and the other listens on 15090
+	// Expect 2 active listeners, one listens on 443 and the other listens on 15090
 	err = ingressutil.WaitUntilGatewayActiveListenerStatsGE(t, ing, 2, 60*time.Second)
 	if err != nil {
 		t.Errorf("total active listener stats does not match: %v", err)
 	}
 	tlsContext := ingressutil.TLSContext{
-		CaCert: ingressutil.CaCertA,
+		CaCert:     ingressutil.CaCertA,
+		PrivateKey: ingressutil.TLSClientKeyA,
+		Cert:       ingressutil.TLSClientCertA,
 	}
-	callType := ingress.TLS
+	callType := ingress.Mtls
 
-	for _, h := range hosts {
+	for _, h := range multipleTLSHosts {
 		err := ingressutil.VisitProductPage(ing, h, callType, tlsContext, 30*time.Second,
 			ingressutil.ExpectedResponse{ResponseCode: 200, ErrorMessage: ""}, t)
 		if err != nil {
@@ -64,11 +62,11 @@ func testMultiTLSGateways(t *testing.T, ctx framework.TestContext) { // nolint:i
 	}
 }
 
-func TestTlsGateways(t *testing.T) {
+func TestMtlsGateways(t *testing.T) {
 	framework.
 		NewTest(t).
 		RequiresEnvironment(environment.Kube).
 		Run(func(ctx framework.TestContext) {
-			testMultiTLSGateways(t, ctx)
+			testMultiMtlsGateways(t, ctx)
 		})
 }
