+++ tests/integration/security/sds_ingress/single_mtls_gateway_separate_secret_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package singlemtlsgatewayseparatesecret
+package sdsingress
 
 import (
 	"testing"
@@ -22,12 +22,6 @@ import (
 	"istio.io/istio/pkg/test/framework/components/environment"
 	"istio.io/istio/pkg/test/framework/components/ingress"
 	ingressutil "istio.io/istio/tests/integration/security/sds_ingress/util"
-)
-
-var (
-	credName   = []string{"bookinfo-credential-1"}
-	credCaName = []string{"bookinfo-credential-1-cacert"}
-	host       = "bookinfo1.example.com"
 )
 
 // TestSingleMTLSGateway_ServerKeyCertRotation tests a single mTLS ingress gateway with SDS enabled.
@@ -47,8 +41,8 @@ func TestSingleMTLSGateway_ServerKeyCertRotation(t *testing.T) {
 			ingressutil.DeployBookinfo(t, ctx, g, ingressutil.SingleMTLSGateway)
 
 			// Add two kubernetes secrets to provision server key/cert and client CA cert for ingress gateway.
-			ingressutil.CreateIngressKubeSecret(t, ctx, credCaName, ingress.Mtls, ingressutil.IngressCredentialCaCertA)
-			ingressutil.CreateIngressKubeSecret(t, ctx, credName, ingress.Mtls,
+			ingressutil.CreateIngressKubeSecret(t, ctx, singleTLCredCAName, ingress.Mtls, ingressutil.IngressCredentialCaCertA)
+			ingressutil.CreateIngressKubeSecret(t, ctx, singleTLSCredName, ingress.Mtls,
 				ingressutil.IngressCredentialServerKeyCertA)
 
 			ingA := ingress.NewOrFail(t, ctx, ingress.Config{Istio: inst})
@@ -67,40 +61,40 @@ func TestSingleMTLSGateway_ServerKeyCertRotation(t *testing.T) {
 				PrivateKey: ingressutil.TLSClientKeyA,
 				Cert:       ingressutil.TLSClientCertA,
 			}
-			err = ingressutil.VisitProductPage(ingA, host, ingress.Mtls, tlsContext, 30*time.Second,
+			err = ingressutil.VisitProductPage(ingA, singleTLSHost, ingress.Mtls, tlsContext, 30*time.Second,
 				ingressutil.ExpectedResponse{ResponseCode: 200, ErrorMessage: ""}, t)
 			if err != nil {
-				t.Errorf("unable to retrieve code 200 from product page at host %s: %v", host, err)
+				t.Errorf("unable to retrieve code 200 from product page at host %s: %v", singleTLSHost, err)
 			}
 
 			// key/cert rotation using mis-matched server key/cert. The server cert cannot pass validation
 			// at client side.
-			ingressutil.RotateSecrets(t, ctx, credName, ingress.Mtls, ingressutil.IngressCredentialServerKeyCertB)
+			ingressutil.RotateSecrets(t, ctx, singleTLSCredName, ingress.Mtls, ingressutil.IngressCredentialServerKeyCertB)
 			// Expect 1 more SDS updates for the server key/cert update.
 			err = ingressutil.WaitUntilGatewaySdsStatsGE(t, ingA, 3, 30*time.Second)
 			if err != nil {
 				t.Errorf("sds update stats does not match: %v", err)
 			}
 			// Client uses old server CA cert to set up SSL connection would fail.
-			err = ingressutil.VisitProductPage(ingA, host, ingress.Mtls, tlsContext, 30*time.Second,
+			err = ingressutil.VisitProductPage(ingA, singleTLSHost, ingress.Mtls, tlsContext, 30*time.Second,
 				ingressutil.ExpectedResponse{ResponseCode: 0, ErrorMessage: "certificate signed by unknown authority"}, t)
 			if err != nil {
-				t.Errorf("unable to retrieve 0 from product page at host %s: %v", host, err)
+				t.Errorf("unable to retrieve 0 from product page at host %s: %v", singleTLSHost, err)
 			}
 
 			// key/cert rotation using matched server key/cert. This time the server cert is able to pass
 			// validation at client side.
-			ingressutil.RotateSecrets(t, ctx, credName, ingress.Mtls, ingressutil.IngressCredentialServerKeyCertA)
+			ingressutil.RotateSecrets(t, ctx, singleTLSCredName, ingress.Mtls, ingressutil.IngressCredentialServerKeyCertA)
 			// Expect 1 more SDS updates for the server key/cert update.
 			err = ingressutil.WaitUntilGatewaySdsStatsGE(t, ingA, 4, 30*time.Second)
 			if err != nil {
 				t.Errorf("sds update stats does not match: %v", err)
 			}
 			// Use old CA cert to set up SSL connection would succeed this time.
-			err = ingressutil.VisitProductPage(ingA, host, ingress.Mtls, tlsContext, 30*time.Second,
+			err = ingressutil.VisitProductPage(ingA, singleTLSHost, ingress.Mtls, tlsContext, 30*time.Second,
 				ingressutil.ExpectedResponse{ResponseCode: 200, ErrorMessage: ""}, t)
 			if err != nil {
-				t.Errorf("unable to retrieve code 200 from product page at host %s: %v", host, err)
+				t.Errorf("unable to retrieve code 200 from product page at host %s: %v", singleTLSHost, err)
 			}
 		})
 }
